refactoring : 파일 분리 #3
