+++ app/src/main/java/org/android/go/sopt/presentation/main/MainActivity.kt
@@ -5,6 +5,9 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import org.android.go.sopt.R
 import org.android.go.sopt.databinding.ActivityMainBinding
+import org.android.go.sopt.presentation.main.gallery.GalleryFragment
+import org.android.go.sopt.presentation.main.home.HomeFragment
+import org.android.go.sopt.presentation.main.search.SearchFragment
 
 
 class MainActivity : AppCompatActivity() {
