+++ app/src/main/java/org/android/go/sopt/presentation/main/search/SearchFragment.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.presentation.main
+package org.android.go.sopt.presentation.main.search
 
 import android.content.Context
 import android.os.Bundle
