+++ app/src/main/res/layout/fragment_gallery.xml
@@ -3,12 +3,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.main.GalleryFragment">
+    tools:context=".presentation.main.gallery.GalleryFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:text="갤러리 뷰" />
 
 </FrameLayout>
