+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingBottom="56dp"
-    tools:context=".presentation.main.HomeFragment">
+    tools:context=".presentation.main.home.HomeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_home_dog"
